Ingest to Qdrant/requirements.txt
quixstreams==2.5.1
python-dotenv
qdrant-client